requirements.txt
selenium
pillow
pytesseract
opencv-python-headless
webdriver-manager
